requirements.txt
fastapi[all]==0.111.0
uvicorn==0.29.0
boto3==1.34.105
pytest==8.2.0
pytest-cov==5.0.0
black==24.4.2
pydantic-settings==2.2.1
motor==3.4.0
python-jose==3.3.0
jwt==1.3.1
httpx==0.23.0
pytest-asyncio==0.23.7
pillow==11.2.1
pymongo==4.6.1